Make space a BigIntegerField

+++ main/models.py
@@ -32,7 +32,7 @@ def generate_image_id() -> str:
 class User(AbstractUser):
     upgraded_until = models.DateField(default=datetime.date(1900, 1, 1))
     last_payment = models.DateField(default=datetime.date(1900, 1, 1))
-    storage_space = models.PositiveIntegerField(default=0)
+    storage_space = models.BigIntegerField(default=0)
     api_key = models.CharField(
         max_length=200, default=generate_moderate_id, unique=True
     )
